refactor(render): move to graphics, add clear
